update v 1.0.0 : modify main

+++ domain/Data.go
@@ -17,6 +17,6 @@ type DatosSensor struct {
 	} `json:"mlx90614"`
 	
 	GSR struct{
-		Porcentaje int
+		Porcentaje float64
 	}`json:"GSR"`	
 }
